refactor classes to use inheritance, answer questions
