+++ student.rb
@@ -1,15 +1,11 @@
-class Student
+require_relative 'pie'
+
+class Student < Pie
   attr_reader :age, :phase
   attr_accessor :name
-
   def initialize(options = {})
-    @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
+    super
+     @phase = options.fetch(:phase, 1)
   end
 
   def set_phase(num)
